Update usePaymentHistory.ts to parse total as float instead of int to preserve decimals (199.50) in payment history display.

X-Lovable-Edit-ID: edt-86a1fe5d-201f-467e-aba0-dc2f080edd8a

+++ src/hooks/usePaymentHistory.ts
@@ -22,7 +22,7 @@ export const usePaymentHistory = () => {
       transaction_id: item.txn_id || item.transaction_id || "",
       amount:
         typeof item.Total === "string"
-          ? parseInt(item.Total) * 100
+          ? parseFloat(item.Total) * 100
           : (item.Total || 0) * 100, // Convert to cents
       currency: item.currancy || item.currency || "usd",
       date: item.CreateTime || item.date || new Date().toISOString(),
